+++ controllers/admin/AdminItellashippingItellaManifestDone.php
@@ -4,6 +4,7 @@ require_once _PS_MODULE_DIR_ . 'itellashipping/vendor/itella-api/vendor/autoload
 use PrestaShop\PrestaShop\Adapter\Entity\PrestaShopException;
 use Mijora\Itella\Pdf\Manifest;
 use Mijora\Itella\CallCourier;
+use Mijora\Itella\ItellaException;
 
 class AdminItellashippingItellaManifestDoneController extends ModuleAdminController
 {
@@ -72,10 +73,14 @@ class AdminItellashippingItellaManifestDoneController extends ModuleAdminControl
     $storeObj = new ItellaStore();
     $stores = ItellaStore::getStores();
 
+    $call_message = Configuration::get('ITELLA_CALL_MESSAGE');
+    if (!$call_message) $call_message = '';
+
     $this->context->smarty->assign(array(
       'moduleMenu' => $menu,
       'stores' => json_encode($stores),
       'call_url' => $this->context->link->getAdminLink($this->controller_name, true),
+      'call_message' => $call_message
     ));
 
     return $this->context->smarty->fetch(_PS_MODULE_DIR_ . 'itellashipping/views/templates/admin/manifest_menu.tpl');
@@ -160,17 +165,83 @@ class AdminItellashippingItellaManifestDoneController extends ModuleAdminControl
         echo json_encode(array('error' => $this->l('Missing manifest ID')));
         exit();
       }
-      $manifest = new ItellaManifest(Tools::getValue('id_manifest'));
-
+      
       ItellaShipping::checkForClass('ItellaStore');
       $storeObj = new ItellaStore(Tools::getValue('id_itella_store'));
+      if (!Validate::isLoadedObject($storeObj)) {
+        echo json_encode(array('error' => $this->l('Failed to retrieve store data. Check the "Stores" page in the module settings.')));
+        exit();
+      }
+      ItellaShipping::checkForClass('ItellaManifest');
+      $manifest = new ItellaManifest(Tools::getValue('id_manifest'));
+      if (!Validate::isLoadedObject($manifest)) {
+        echo json_encode(array('error' => $this->l('Failed to retrieve manifest data')));
+        exit();
+      }
+
       $isTest = (Configuration::get('ITELLA_TEST_MODE') == 1);
-      
-      $mail_sent = false;
-      
+      $username = Configuration::get('ITELLA_API_USER_2317');
+      $password = Configuration::get('ITELLA_API_PASS_2317');
+      if (empty($username)) {
+        $username = Configuration::get('ITELLA_API_USER_2711');
+        $password = Configuration::get('ITELLA_API_PASS_2711');
+      }
+      $email_send_to = Configuration::get('ITELLA_CALL_EMAIL_' . strtoupper($storeObj->country_code));
+      $email_subject = Configuration::get('ITELLA_CALL_EMAIL_SUBJECT');
+      $sender_name = Configuration::get('ITELLA_SENDER_NAME');
+      $sender_code = Configuration::get('ITELLA_SENDER_CODE');
+
+      $call_message = Tools::getValue('call_message');
+      $call_date = Tools::getValue('call_date');
+      $call_time_from = Tools::getValue('call_time_from');
+      $call_time_to = Tools::getValue('call_time_to');
+
+      $errors = array();
+      $success = array();
+
       try {
-        $send_to = Configuration::get('ITELLA_CALL_EMAIL_' . strtoupper($storeObj->country_code));
-        if (!$send_to) {
+        $caller = new CallCourier($email_send_to, $isTest);
+        $caller
+          ->setUsername($username)
+          ->setPassword($password)
+          ->setSenderEmail($this->context->employee->email)
+          ->setSubject($email_subject)
+          ->setPickUpAddress(array(
+            'sender' => $sender_name,
+            'address_1' => $storeObj->address,
+            'postcode' => $storeObj->postcode,
+            'city' => $storeObj->city,
+            'country' => strtoupper($storeObj->country_code),
+            'contact_phone' => $storeObj->phone,
+          ))
+          ->setPickUpParams(array(
+            'date' => $call_date,
+            'time_from' => $call_time_from,
+            'time_to' => $call_time_to,
+            'info_general' => $call_message,
+            'id_sender' => $sender_code,
+          ))
+          ->setAttachment($manifest->getManifestString(), true)
+          ->setItems($manifest->getManifestItems())
+          ->disableMethod('email');
+
+        $result_api = $caller->callCourier();
+
+        if (!empty($result_api['errors'])) {
+          $errors = array_merge($errors, $result_api['errors']);
+        }
+        if (!empty($result_api['success'])) {
+          $success = array_merge($success, $result_api['success']);
+        }
+        if (empty($result_api['errors']) && empty($result_api['success'])) {
+          $errors[] = $this->l('An unknown result was received from API library');
+        }
+      } catch (\ItellaException $e) {
+        $errors[] = $this->l('Call courier failed with:') . ' ' . $e->getMessage();
+      }
+
+      try {
+        if (!$email_send_to) {
           throw new \Exception($this->l('Courier service email not set'));
         }
 
@@ -187,64 +258,95 @@ class AdminItellashippingItellaManifestDoneController extends ModuleAdminControl
 
         $dir_mail = $this->module->getLocalPath() . 'mails/';
 
-        $mailer = new CallCourier($send_to, $isTest);
-        $body_html = $mailer
-          ->setSenderEmail($this->context->employee->email)
-          ->setSubject(Configuration::get('ITELLA_CALL_EMAIL_SUBJECT'))
-          ->setPickUpAddress(array(
-            'sender' => Configuration::get('ITELLA_SENDER_NAME'),
-            'address_1' => $storeObj->address,
-            'postcode' => $storeObj->postcode,
-            'city' => $storeObj->city,
-            'country' => $storeObj->country_code,
-            //'pickup_time' => $storeObj->pick_start . ' - ' . $storeObj->pick_finish,
-            'contact_phone' => $storeObj->phone,
-          ))
-          ->setAttachment(true)
-          ->setItems($manifest->getManifestItems())
-          ->buildMailBody();
-
-        try {
-          $mailer->callApiCourier();
-        } catch (\Exception $e) {
-          // Ignore this
-        }
-
         $data = array(
-          '{title}' => Configuration::get('ITELLA_CALL_EMAIL_SUBJECT'),
-          '{body}' => $body_html
+          '{title}' => $email_subject,
+          '{body}' => $caller->buildMailBody()
         );
 
         $mail_sent = Mail::Send(
           $id_lang,
           'itella_call_courier',
-          ($isTest ? 'TEST CALL - ' : '') . Configuration::get('ITELLA_CALL_EMAIL_SUBJECT'),
+          ($isTest ? 'TEST CALL - ' : '') . $email_subject,
           $data,
-          $send_to,
+          $email_send_to,
           'Smartposti Courier Service',
           $this->context->employee->email,
-          Configuration::get('ITELLA_SENDER_NAME'),
+          $sender_name,
           $file_attachement,
           null,
           $dir_mail,
           false,
           (int) $storeObj->id_shop
         );
-      } catch (\Exception $th) {
-        echo json_encode(array('error' => $this->l('Call courier failed with:') . ' ' . $th->getMessage()));
-        exit();
+
+        $success[] = $this->l('Courier call email sent');
+      } catch (\Exception $e) {
+        $errors[] = $this->l('Failed to send courier call email') . ': ' . $e->getMessage();
       }
 
-      if (!$mail_sent) {
-        echo json_encode(array('error' => $this->l('Call courier failed')));
-        exit();
+      $return = array();
+      if (empty($success) && !empty($errors)) {
+        $return['error'] = implode('<br/>', $errors);
+      } else if (!empty($errors)) {
+        $return['success'] = implode('<br/>', $success) . '<br/><br/><span style="color:#f55;">' . implode('<br/>', $errors) . '</span>';
+      } else if (!empty($success)) {
+        $return['success'] = implode('<br/>', $success);
+      } else {
+        $return['error'] = $this->l('An unknown result was received from the call');
       }
 
-      echo json_encode(array('success' => $this->l('Smartposti courier called')));
+      echo json_encode($return);
       exit();
     }
     echo json_encode(array('error' => $this->l('Bad store address ID')));
     exit();
+  }
+
+  private function sendCallEmail($caller, $send_to, $manifest_string, $shop_id)
+  {
+    $id_lang = Language::getIdByIso('en');
+    if (!$id_lang) {
+      $id_lang = (int)Configuration::get('PS_LANG_DEFAULT');
+    }
+
+    $isTest = (Configuration::get('ITELLA_TEST_MODE') == 1);
+
+    $file_attachement = array(
+      'content' => $manifest_string,
+      'name' => 'manifest.pdf',
+      'mime' => 'application/pdf'
+    );
+    $dir_mail = $this->module->getLocalPath() . 'mails/';
+
+    try {
+      $data = array(
+        '{title}' => Configuration::get('ITELLA_CALL_EMAIL_SUBJECT'),
+        '{body}' => $caller->buildMailBody()
+      );
+
+      $mail_sent = Mail::Send(
+        $id_lang,
+        'itella_call_courier',
+        ($isTest ? 'TEST CALL - ' : '') . Configuration::get('ITELLA_CALL_EMAIL_SUBJECT'),
+        $data,
+        $send_to,
+        'Smartposti Courier Service',
+        $this->context->employee->email,
+        Configuration::get('ITELLA_SENDER_NAME'),
+        $file_attachement,
+        null,
+        $dir_mail,
+        false,
+        (int) $shop_id
+      );
+      if (!$mail_sent) {
+        return array('error' => $this->l('Call courier failed'));
+      }
+    } catch (\Exception $e) {
+      return array('error' => $this->l('Call courier failed with:') . ' ' . $e->getMessage());
+    }
+
+    return array('success' => $this->l('Courier called successfully'));
   }
 
   public function renderList()